Move all managed SNI classes to ManagedSni namespace
