+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/ManagedSni/LocalDB.netcore.Windows.cs
@@ -9,8 +9,7 @@ using Interop.Windows.Handles;
 using Interop.Windows.Kernel32;
 using Microsoft.Win32;
 
-
-namespace Microsoft.Data.SqlClient.SNI
+namespace Microsoft.Data.SqlClient.ManagedSni
 {
     internal sealed class LocalDB
     {
