+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/ManagedSni/SNIError.netcore.cs
@@ -4,7 +4,7 @@
 
 using System;
 
-namespace Microsoft.Data.SqlClient.SNI
+namespace Microsoft.Data.SqlClient.ManagedSni
 {
     /// <summary>
     /// SNI error
