+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/ManagedSni/SNIMarsHandle.netcore.cs
@@ -7,7 +7,7 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.Threading;
 
-namespace Microsoft.Data.SqlClient.SNI
+namespace Microsoft.Data.SqlClient.ManagedSni
 {
     /// <summary>
     /// MARS handle
