+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/ManagedSni/SNITcpHandle.netcore.cs
@@ -15,7 +15,7 @@ using System.Threading;
 using Microsoft.Data.Common;
 using Microsoft.Data.ProviderBase;
 
-namespace Microsoft.Data.SqlClient.SNI
+namespace Microsoft.Data.SqlClient.ManagedSni
 {
     /// <summary>
     /// TCP connection handle
